+++ CHANGELOG.md
@@ -6,9 +6,26 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Added
+
+- starknet.subsquid.events: Added `starknet.subsquid.events` index kind to process Starknet events.
+- starknet.subsquid: Added `starknet.subsquid` datasource to fetch historical data from Subsquid Archives.
+
+### Fixed
+
+- config: Fixed setting logging levels according to the config.
+- evm.events: Fixed matching logs when filtering by topic0.
+
 ### Other
 
 - deps: `tortoise-orm` updated to 0.21.2.
+
+## [7.5.7] - 2024-05-30
+
+### Fixed
+
+- config: Fixed setting logging levels according to the config.
+- evm.subsquid.events: Fixed matching logs when filtering by topic0.
 
 ## [7.5.6] - 2024-05-16
 
@@ -1472,7 +1489,11 @@ This release contains no changes except for the version number.
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/dipdup-io/dipdup/compare/7.5.4...HEAD
+[Unreleased]: https://github.com/dipdup-io/dipdup/compare/8.0.0a1...HEAD
+[7.5.7]: https://github.com/dipdup-io/dipdup/compare/7.5.6...7.5.7
+[7.5.6]: https://github.com/dipdup-io/dipdup/compare/7.5.5...7.5.6
+[8.0.0a1]: https://github.com/dipdup-io/dipdup/compare/7.5.7...8.0.0a1
+[7.5.5]: https://github.com/dipdup-io/dipdup/compare/7.5.4...7.5.5
 [7.5.4]: https://github.com/dipdup-io/dipdup/compare/7.5.3...7.5.4
 [7.5.3]: https://github.com/dipdup-io/dipdup/compare/7.5.2...7.5.3
 [7.5.2]: https://github.com/dipdup-io/dipdup/compare/7.5.1...7.5.2